Only cache logos for validators that are set to active
There are over 700 validators in the validator list, this causes extreme lag in pulling logos/avatars for validators. Validators that have their active flag set to true, should be the only one's we're pulling logos for, this will help tremendously with validators complaining their logos aren't showing up. It's not a permanaent fix for this, as I believe this module needs to be re-written but it's better than the current state